refactor plugin installation to plugins folder
